job_automation/requirements.txt
requests
beautifulsoup4
openai
streamlit==1.32.0
plotly
google-auth-oauthlib
google-api-python-client
